feat(model-server): isolated storage of repository data
The versions of models where previously stored in the same storage
without knowing to which repository they belong.
To allow proper permission checks there is now an additional
`repository` field in the database.

Existing repositories remain in the global storage and don't benefit
from this new security feature. A migration is not yet implemented.

New repositories are created in the new storage, but are not accessible
via the deprecated `RestWebModelClient`.